Final fixes and Javadoc documentation

+++ src/main/java/lanej/inventorysystem/controller/ModifyProduct.java
@@ -15,37 +15,78 @@ import java.net.URL;
 import java.util.Comparator;
 import java.util.ResourceBundle;
 
+/**
+ * The ModifyProduct class is responsible for managing the user interface and functionality of the Modify Product
+ * screen in the inventory management system. This class implements the Initializable interface from the JavaFX API.
+ * It contains text fields, tables, and buttons for modifying the details of an existing product, including its
+ * associated parts. The class allows users to search, add, and remove parts from the associated parts table, as well
+ * as save and cancel any modifications made to the product's details.
+ * @author Jonathan Lane
+ */
 public class ModifyProduct implements Initializable {
-
+    /** Displays the Product's ID, which is taken from the Product being modified */
     public TextField idField;
+    /** Displays the Product's name, editable by the user */
     public TextField nameField;
+    /** Displays the Product's stock, editable by the user */
     public TextField inventoryField;
+    /** Displays the Product's price, editable by the user */
     public TextField priceField;
+    /** Displays the Product's maximum stock, editable by the user */
     public TextField maxField;
+    /** Displays the Product's minimum stock, editable by the user */
     public TextField minField;
+    /** A TableView of all the available Parts within Inventory */
     public TableView partTable;
+    /** The Part ID column for partTable */
     public TableColumn partIdColumn;
+    /** The Part name column for partTable */
     public TableColumn partNameColumn;
+    /** The Part stock column for partTable */
     public TableColumn partInventoryColumn;
+    /** The Part price column for partTable */
     public TableColumn partPriceColumn;
+    /** The Part max column for partTable */
     public TableColumn partMaxColumn;
+    /** The Part min column for partTable */
     public TableColumn partMinColumn;
+    /** The search field for filtering partTable */
     public TextField partSearchField;
+    /** A TableView of all the Parts associated with the Product */
     public TableView associatedTable;
+    /** The Part ID column for associatedTable */
     public TableColumn associatedIdColumn;
+    /** The Part name column for associatedTable */
     public TableColumn associatedNameColumn;
+    /** The Part stock column for associatedTable */
     public TableColumn associatedInventoryColumn;
+    /** The Part price column for associatedTable */
     public TableColumn associatedPriceColumn;
+    /** The Part max column for associatedTable */
     public TableColumn associatedMaxColumn;
+    /** The Part min column for associatedTable */
     public TableColumn associatedMinColumn;
+    /** The search field for filtering associatedTable */
     public TextField associatedSearchField;
 
-    // We still need productToAdd in the case that the user clicks the Cancel button
+    /** The Product that is being added. It's necessary as a buffer in case the user decides to cancel the modification. */
     private final Product productToAdd = new Product(9999,"",0.0,0,0,0); // Must change later
+    /** The Product that is being referenced, it is assigned the Product that is selected from the Main Screen. */
     private Product productToModify = null;
+    /** The FilteredList of all available parts, which allows us to filter the items using the partSearchField */
     private FilteredList<Part> partFilteredList;
+    /** The FilteredList of all associated parts, which allows us to filter the items using the associatedSearchField */
     private FilteredList<Part> associatedFilteredList;
 
+    /**
+     * This method is called by the FXMLLoader when the modify-product.fxml file is loaded. It sets the initial values
+     * for all the fields in the screen, populates the parts table, and associated parts table with the parts that
+     * are already associated with the product that is being modified. It also initializes the sorting and filtering
+     * functionality for both the all parts table and the associated parts table. If the product being modified no
+     * longer exists, an error message is displayed and the user is redirected to the main screen.
+     * @param url The URL used to initialize the controller.
+     * @param resourceBundle The ResourceBundle used to initialize the controller.
+     */
     @Override
     public void initialize(URL url, ResourceBundle resourceBundle) {
         System.out.println("ModifyProduct Controller Initialized");
@@ -133,6 +174,12 @@ public class ModifyProduct implements Initializable {
         }));
     }
 
+    /**
+     * This initializes and sorts the associatedFilteredList, and populates the associatedTable with it. This simplifies
+     * my attempts to sort the associatedTable, as it would not automatically sort itself when a new Part association
+     * is added. So to make it sort every time, I manually recreate the ObservableList that gets sorted before being
+     * added to the associatedFilteredList every time that a Part association is added.
+     */
     private void initSortedAssocParts() {
         ObservableList<Part> sortedAssociated = productToAdd.getAllAssociatedParts();
         sortedAssociated.sort(Comparator.comparingInt(Part::getId));
@@ -142,6 +189,12 @@ public class ModifyProduct implements Initializable {
         associatedTable.sort();
     }
 
+    /**
+     * This method is called when the user presses the "Add Part Association" button. It adds the Part association to
+     * the Product object that would be added to Inventory if the user clicks "Save" afterward. If the same Part is
+     * already associated with the Product for whatever reason, it asks the user to confirm before adding the duplicate
+     * Part.
+     */
     public void addButton() {
         if (partTable.getSelectionModel().getSelectedItem() != null) {
             // Get selected Part
@@ -161,12 +214,33 @@ public class ModifyProduct implements Initializable {
         }
     }
 
+    /**
+     * This method is called whenever the user clicks the "Remove Part Association" button. It will ask the user for
+     * confirmation before calling Product.deleteAssociatedPart(productToAdd). It only does this if the user has
+     * selected an item within the Associated Parts table, otherwise it doesn't do anything.
+     */
     public void removeButton() {
         if (associatedTable.getSelectionModel().getSelectedItem() != null) {
-            productToAdd.deleteAssociatedPart((Part)associatedTable.getSelectionModel().getSelectedItem());
+            Alert removeConfirmation = new Alert(Alert.AlertType.CONFIRMATION, "Are you sure you want to remove " +
+                    ((Part)associatedTable.getSelectionModel().getSelectedItem()).getName() + "?");
+            removeConfirmation.showAndWait();
+            if (removeConfirmation.getResult() == ButtonType.OK) {
+                productToAdd.deleteAssociatedPart((Part)associatedTable.getSelectionModel().getSelectedItem());
+            }
         }
     }
 
+    /**
+     * This method is called whenever the user clicks the "Save" button. There exist many try and catch blocks within
+     * this method due to the need to check and notify the user of exactly what is wrong with their input, if anything.
+     * Ultimately, this method attempts to obtain the user input for all the editable TextFields within this screen
+     * (not counting the search fields), and parse that information into the data for the creation of a new Product
+     * (that is then added to Inventory).
+     * Since this is the Modify Product screen, it will also remove the old Product from Inventory. It gets the new
+     * Product's ID from the old Product. Upon successful modification (technically recreation), it will direct the
+     * user back to the Main Screen.
+     * @param event The ActionEvent that is created from the Save button within the scene
+     */
     public void saveButton(ActionEvent event) {
         try {
             String name = nameField.getText();
@@ -240,7 +314,14 @@ public class ModifyProduct implements Initializable {
         InventoryApplication.toScreen(event, InventoryApplication.ScreenType.MAIN_SCREEN);
     }
 
+    /**
+     * This method is called whenever the user clicks the "Cancel" button. It simply directs the user back to the Main
+     * Screen, without saving any info that may have been created in the other actions of the scene. It also sets the
+     * InventoryApplication's static productToModify object to null before returning to ensure consistency.
+     * @param event The ActionEvent that is created from the Cancel button within the scene
+     */
     public void cancelButton(ActionEvent event) {
+        InventoryApplication.productToModify = null;
         InventoryApplication.toScreen(event, InventoryApplication.ScreenType.MAIN_SCREEN);
     }
 }
